requirements.txt
alembic==1.4.2
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
netifaces==0.10.6
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
six==1.14.0
socketio==0.2.1
SQLAlchemy==1.3.16
Werkzeug==1.0.1
